Implemented AV2-181: Implement validate feature -validation refactoring
