+++ etc/di.xml
@@ -15,6 +15,9 @@
         </arguments>
     </type>
     <type name="Magento\Checkout\Model\ShippingInformationManagement">
-        <plugin name="onepica_avatax_addressvalidation" type="OnePica\AvaTax\Plugin\Checkout\Model\ShippingInformationManagement" />
+        <plugin name="onepica_avatax_onepage_address_validation" type="OnePica\AvaTax\Plugin\Checkout\Model\ShippingInformationManagement" />
+    </type>
+    <type name="Magento\Quote\Model\Quote\Address">
+        <plugin name="quote_model_quote_address_validator" type="OnePica\AvaTax\Plugin\Quote\Model\Quote\AddressValidator" />
     </type>
 </config>
